Seperate de logic and the display of Searchar component and MovieList component
