--- client/src/components/MovieList/searchBar.tsx
@@ -1,31 +1,0 @@
-type SearchBarProps = {
-  searchTerm: string;
-  setSearchTerm: (value: string) => void;
-};
-
-export default function SearchBar({
-  searchTerm,
-  setSearchTerm,
-}: SearchBarProps) {
-  const handleSearchTerm = (event: React.ChangeEvent<HTMLInputElement>) => {
-    const value = event.target.value;
-    setSearchTerm(value);
-  };
-  console.info(searchTerm);
-
-  return (
-    <>
-      <header>
-        <h1>Les Films</h1>
-        <label htmlFor="searchBar">Recherche</label>
-        <input
-          type="search"
-          name="searchBar"
-          className="rechercher"
-          placeholder="Que voulez vous regardez"
-          onChange={handleSearchTerm}
-        />
-      </header>
-    </>
-  );
-}
